+++ R/betaMC-beta-ci-dot.R
@@ -1,8 +1,13 @@
-#' Confidence Intervals for Standardized Regression Coefficients
+#' Confidence Intervals for
+#' Standardized Regression Coefficients
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
-#' @return Returns a matrix.
+#' @return Returns a matrix of
+#'   standardized regression slopes,
+#'   standard errors,
+#'   and
+#'   confidence intervals.
 #'
 #' @param object Object of class `betamc`.
 #' @param alpha Numeric vector.