+++ R/betaMC-pcci-dot.R
@@ -2,6 +2,12 @@
 #'
 #' @author Ivan Jacob Agaloos Pesigan
 #'
+#' @return Returns a matrix of estimates,
+#'   standard errors,
+#'   number of Monte Carlo replications,
+#'   and
+#'   confidence intervals.
+#'
 #' @param thetahatstar Numeric vector.
 #'   Monte Carlo sampling distribution.
 #' @param thetahat Numeric.
@@ -9,9 +15,7 @@
 #' @param alpha Numeric vector.
 #'   Significance level.
 #'   Default value is `alpha = c(0.001, 0.01, 0.05)`.
-#' @return Returns a matrix of estimates,
-#'   standard errors, number of Monte Carlo replications,
-#'   and confidence intervals.
+#'
 #' @keywords internal
 #' @noRd
 .PCCI <- function(thetahatstar,