+++ tests/testthat/test-betaMC-r-sq-beta-mc.R
@@ -1,4 +1,4 @@
-## ---- test-betaMC-diff-beta-mc
+## ---- test-betaMC-r-sq-beta-mc
 lapply(
   X = 1,
   FUN = function(i,
@@ -17,15 +17,17 @@ lapply(
     }
     df <- nas1982
     object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = df)
-    mvn <- DiffBetaMC(BetaMC(object, type = "mvn", R = R))
-    adf <- DiffBetaMC(BetaMC(object, type = "adf", R = R))
-    hc0 <- DiffBetaMC(BetaMC(object, type = "hc0", R = R))
-    hc1 <- DiffBetaMC(BetaMC(object, type = "hc1", R = R))
-    hc2 <- DiffBetaMC(BetaMC(object, type = "hc2", R = R))
-    hc3 <- DiffBetaMC(BetaMC(object, type = "hc3", R = R))
-    hc4 <- DiffBetaMC(BetaMC(object, type = "hc4", R = R))
-    hc4m <- DiffBetaMC(BetaMC(object, type = "hc4m", R = R))
-    hc5 <- DiffBetaMC(BetaMC(object, type = "hc5", R = R))
+    r_sq <- summary(object)$r.squared
+    adj <- summary(object)$adj.r.squared
+    mvn <- RSqBetaMC(BetaMC(object, type = "mvn", R = R))
+    adf <- RSqBetaMC(BetaMC(object, type = "adf", R = R))
+    hc0 <- RSqBetaMC(BetaMC(object, type = "hc0", R = R))
+    hc1 <- RSqBetaMC(BetaMC(object, type = "hc1", R = R))
+    hc2 <- RSqBetaMC(BetaMC(object, type = "hc2", R = R))
+    hc3 <- RSqBetaMC(BetaMC(object, type = "hc3", R = R))
+    hc4 <- RSqBetaMC(BetaMC(object, type = "hc4", R = R))
+    hc4m <- RSqBetaMC(BetaMC(object, type = "hc4m", R = R))
+    hc5 <- RSqBetaMC(BetaMC(object, type = "hc5", R = R))
     testthat::test_that(
       paste(text, "mvn", "multiple regression"),
       {
@@ -33,9 +35,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(mvn)
             ) <= tol
           )
@@ -56,9 +57,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(adf)
             ) <= tol
           )
@@ -79,9 +79,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc0)
             ) <= tol
           )
@@ -102,9 +101,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc1)
             ) <= tol
           )
@@ -125,9 +123,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc2)
             ) <= tol
           )
@@ -148,9 +145,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc3)
             ) <= tol
           )
@@ -171,9 +167,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc4)
             ) <= tol
           )
@@ -194,9 +189,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc4m)
             ) <= tol
           )
@@ -217,9 +211,8 @@ lapply(
           all(
             abs(
               c(
-                0.4951 - 0.3915,
-                0.4951 - 0.2632,
-                0.3915 - 0.2632
+                r_sq,
+                adj
               ) - coef(hc5)
             ) <= tol
           )
@@ -234,80 +227,217 @@ lapply(
       }
     )
     object <- lm(QUALITY ~ NARTIC, data = df)
+    r_sq <- summary(object)$r.squared
+    adj <- summary(object)$adj.r.squared
+    mvn <- RSqBetaMC(BetaMC(object, type = "mvn", R = R))
+    adf <- RSqBetaMC(BetaMC(object, type = "adf", R = R))
+    hc0 <- RSqBetaMC(BetaMC(object, type = "hc0", R = R))
+    hc1 <- RSqBetaMC(BetaMC(object, type = "hc1", R = R))
+    hc2 <- RSqBetaMC(BetaMC(object, type = "hc2", R = R))
+    hc3 <- RSqBetaMC(BetaMC(object, type = "hc3", R = R))
+    hc4 <- RSqBetaMC(BetaMC(object, type = "hc4", R = R))
+    hc4m <- RSqBetaMC(BetaMC(object, type = "hc4m", R = R))
+    hc5 <- RSqBetaMC(BetaMC(object, type = "hc5", R = R))
     testthat::test_that(
-      paste(text, "mvn", "simple regression"),
+      paste(text, "mvn", "multiple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "mvn"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(mvn)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(mvn)[, "est"] - coef(mvn)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
-      paste(text, "adf", "simple regression"),
+      paste(text, "adf", "multiple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "adf"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(adf)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(adf)[, "est"] - coef(adf)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc0", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc0"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc0)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc0)[, "est"] - coef(hc0)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc1", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc1"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc1)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc1)[, "est"] - coef(hc1)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc2", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc2"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc2)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc2)[, "est"] - coef(hc2)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc3", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc3"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc3)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc3)[, "est"] - coef(hc3)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc4", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc4"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc4)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc4)[, "est"] - coef(hc4)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc4m", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc4m"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc4m)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc4m)[, "est"] - coef(hc4m)
+            ) <= tol
+          )
         )
       }
     )
     testthat::test_that(
       paste(text, "hc5", "simple regression"),
       {
-        testthat::expect_error(
-          DiffBetaMC(BetaMC(object, type = "hc5"))
+        testthat::expect_true(
+          all(
+            abs(
+              c(
+                r_sq,
+                adj
+              ) - coef(hc5)
+            ) <= tol
+          )
+        )
+        testthat::expect_true(
+          all(
+            abs(
+              summary(hc5)[, "est"] - coef(hc5)
+            ) <= tol
+          )
         )
       }
     )
   },
-  text = "test-betaMC-diff-beta-mc",
+  text = "test-betaMC-r-sq-beta-mc",
   R = 10L,
   tol = 0.0001
 )